# clients/typescript/scripts/openapi.snapshot.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Spatial.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5299/"
    }
  ],
  "paths": {
...
    "/arcgis/admin/services/{service}": {
      "get": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/arcgis/admin/services/{service}/createService": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/arcgis/admin/services/{service}/deleteService": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/arcgis/admin/uploads/{id}/publish": {
      "post": {
        "tags": [
          "Spatial.Host"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}